requirements.txt
python-dotenv
requests
anthropic
openai
telebot
groq
cryptography